+++ spark-frontend/src/screens/Faucet/Faucet.tsx
@@ -7,24 +7,9 @@ import SizedBox from "@components/SizedBox";
 import Text, { TEXT_TYPES } from "@components/Text";
 import TokensFaucetTable from "@screens/Faucet/TokensFaucetTable";
 import { media } from "@src/themes/breakpoints";
+import { useStores } from "@stores";
 
 interface IProps {}
-
-const Root = styled.div`
-  display: flex;
-  flex-direction: column;
-  box-sizing: border-box;
-  padding: 0 16px;
-  width: 100%;
-  min-height: 100%;
-  margin-bottom: 24px;
-  margin-top: 40px;
-  text-align: left;
-
-  @media (min-width: 880px) {
-    margin-top: 56px;
-  }
-`;
 
 const Faucet: React.FC<IProps> = observer(() => {
   const { faucetStore } = useStores();
