+++ spark-frontend/src/screens/Faucet/TokensFaucetTable.tsx
@@ -4,9 +4,9 @@ import { observer } from "mobx-react-lite";
 
 import Chip from "@components/Chip";
 import { Column, Row } from "@components/Flex";
-import MintButtons from "@components/MintButtons";
 import { TableText } from "@components/Table";
 import Text, { TEXT_TYPES, TEXT_TYPES_MAP } from "@components/Text";
+import MintButtons from "@screens/Faucet/MintButtons";
 import { useStores } from "@stores";
 
 interface IProps {}
